// old/ProcureHub.WebApp/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        QueryUsersResponse: {
            id: string;
            email: string;
            firstName: string;
            lastName: string;
            roles: string[];
            enabledAt: null | components["schemas"]["NullableOfDateTime"];
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
            deletedAt: null | components["schemas"]["NullableOfDateTime"];
            department: null | components["schemas"]["QueryUsersDepartment"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}